Form1_2 IsDone

+++ LabsOnC/Lab1/formula.c
@@ -1,6 +1,6 @@
 #include "formula.h"
 
-int calculate(int x)
+double calculate(int x)
 {
     double s;
     s = 0;
